moved methods around
